modularized jest/enzyme unit tests
