+++ spec/client/enzyme.js
@@ -2,5 +2,5 @@ import Enzyme, { configure, shallow, mount, render } from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
 
 configure({ adapter: new Adapter() });
-export { shallow, mount, render };
+export { shallow };
 export default Enzyme;
