+++ test/test.js
@@ -1,95 +1,82 @@
-var chai = require('chai');
-var chaiHttp = require('chai-http');
-var app = require('../server/server.js');
-
-var assert = require('assert');
-const db = require('../db/index.js')
-
-chai.use(chaiHttp)
-var should = require('chai').should()
-
-describe('Item tables should render with the correct item types', function() {
-  describe('Item Details Database', function() {
-    it('should return the correct item detail types', function(done) {
-       chai.request('http://localhost:3001')
-        .get('/api/items/50')
-        .end((err, res) => {
-          var resBody = res.body[0]
-          res.should.have.status(200);
-          resBody.should.be.an('object');
-          resBody.itemId.should.be.a('number');
-          resBody.sizing.should.be.a('string')
-          resBody.material.should.be.a('string')
-          resBody.fit.should.be.a('string')
-          resBody.length.should.be.a('string')
-          resBody.features.should.be.a('string')
-          resBody.neckline.should.be.a('string')
-          resBody.itemStyle.should.be.a('string')
-          resBody.careAndCleaning.should.be.a('string')
-          resBody.TCIN.should.be.a('number')
-          resBody.UPC.should.be.a('number')
-          resBody.DPCI.should.be.a('number')
-          resBody.origin.should.be.a('string')
-          resBody.recycledPolyester.should.be.a('boolean')
-          resBody.fastShipping.should.be.a('boolean')
-          resBody.estimatedShipDimensions.should.be.a('string')
-          resBody.estimatedShipWeight.should.be.a('string')
-          resBody.type.should.equal('Men - Shirts')
-          done()
-        })
-    });
-  });
-})
-
-  describe('Sizing Database', function() {
-      it('should return the correct sizing types', function(done) {
-         chai.request('http://localhost:3001')
-          .get('/api/sizing/50')
-          .end((err, res) => {
-            var firstElementInBody = res.body[0]
-            res.should.have.status(200);
-            res.body.should.be.an('array');
-            firstElementInBody.should.be.an('object')
-            firstElementInBody.size.should.be.a('string');
-            firstElementInBody.neck.should.be.a('string')
-            firstElementInBody.chest.should.be.a('string')
-            done()
-          })
-      });
-  });
-
-
-  describe('Questions Database', function() {
-      it('should return the correct question types', function(done) {
-         chai.request('http://localhost:3001')
-          .get('/api/questions/50')
-          .end((err, res) => {
-            console.log('res', res)
-            res.should.have.status(200);
-            res.body.should.be.an('array');
-            var firstQuestion = res.body[0]
-            firstQuestion.should.be.an('object')
-            firstQuestion.itemId.should.be.a('number');
-            firstQuestion.question.should.be.a('string')
-            firstQuestion.asker.should.be.a('string')
-            done()
-          })
-      });
-  });
-
-
-//   fastShipping: Boolean,
-//   estimatedShipDimensions: String,
-//   estimatedShipWeight: String,
-//   type: men’s shirts
+// var chai = require('chai');
+// var chaiHttp = require('chai-http');
+// var app = require('../server/server.js');
 //
-// await sizing db where type = men’s shirts
+// var assert = require('assert');
+// const db = require('../db/index.js')
 //
-//   size: { type: String },
-//   neck: String,
-//   chest: String,
-//   sleeve: String,
+// chai.use(chaiHttp)
+// var should = require('chai').should()
 //
+// describe('Item tables should render with the correct item types', function() {
+//   describe('Item Details Database', function() {
+//     it('should return the correct item detail types', function(done) {
+//        chai.request('http://localhost:3001')
+//         .get('/api/items/50')
+//         .end((err, res) => {
+//           var resBody = res.body[0]
+//           res.should.have.status(200);
+//           resBody.should.be.an('object');
+//           resBody.itemId.should.be.a('number');
+//           resBody.sizing.should.be.a('string')
+//           resBody.material.should.be.a('string')
+//           resBody.fit.should.be.a('string')
+//           resBody.length.should.be.a('string')
+//           resBody.features.should.be.a('string')
+//           resBody.neckline.should.be.a('string')
+//           resBody.itemStyle.should.be.a('string')
+//           resBody.careAndCleaning.should.be.a('string')
+//           resBody.TCIN.should.be.a('number')
+//           resBody.UPC.should.be.a('number')
+//           resBody.DPCI.should.be.a('number')
+//           resBody.origin.should.be.a('string')
+//           resBody.recycledPolyester.should.be.a('boolean')
+//           resBody.fastShipping.should.be.a('boolean')
+//           resBody.estimatedShipDimensions.should.be.a('string')
+//           resBody.estimatedShipWeight.should.be.a('string')
+//           resBody.type.should.equal('Men - Shirts')
+//           done()
+//         })
+//     });
+//   });
+// })
+//
+//   describe('Sizing Database', function() {
+//       it('should return the correct sizing types', function(done) {
+//          chai.request('http://localhost:3001')
+//           .get('/api/sizing/50')
+//           .end((err, res) => {
+//             var firstElementInBody = res.body[0]
+//             res.should.have.status(200);
+//             res.body.should.be.an('array');
+//             firstElementInBody.should.be.an('object')
+//             firstElementInBody.size.should.be.a('string');
+//             firstElementInBody.neck.should.be.a('string')
+//             firstElementInBody.chest.should.be.a('string')
+//             done()
+//           })
+//       });
+//   });
+//
+//
+//   describe('Questions Database', function() {
+//       it('should return the correct question types', function(done) {
+//          chai.request('http://localhost:3001')
+//           .get('/api/questions/50')
+//           .end((err, res) => {
+//             console.log('res', res)
+//             res.should.have.status(200);
+//             res.body.should.be.an('array');
+//             var firstQuestion = res.body[0]
+//             firstQuestion.should.be.an('object')
+//             firstQuestion.itemId.should.be.a('number');
+//             firstQuestion.question.should.be.a('string')
+//             firstQuestion.asker.should.be.a('string')
+//             done()
+//           })
+//       });
+//   });
+
 //
 // await questions db, grab the first row — check typeof
 // 	  itemId: Number,
